requirements.txt
aiobotocore==2.9.0
aiofiles==23.2.1
aiohttp==3.9.1
fastapi==0.108.0
litellm==1.15.7
numpy==1.26.1
openai==1.3.5
pydantic==2.5
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
tiktoken==0.5.2
twilio==8.9.0
uvicorn==0.22.0
websockets==10.4
onnxruntime==1.16.3
scipy==1.11.4
uvloop==0.19.0sentence-transformers
faiss-cpu
